# sort: inline search convenience wrappers